refactor: stick to old config file to simplify deployments
